Visual indicator as extension badge while searching/handling a CMP.
Added option "noTimeout" to ClickAction, that if used, prevents COM from using even a setTimeout(0) after the click, this is sometimes needed for performance reasons.

Added OnOffMatcher that has a css selector for both on and off state of a consent. This enables us to only click if the state is wrong, saving clicks, and potential performance improvement.

Removed some left over logging, that was not behind the debugLog flag.

Fixed a bug realted to multiple CMP searches running on top of each other.

Added negateChildFilter option, that enables a childfilter to filter out possible targets, based on not having a specific child. (:has, is not supported in all browsers yet)